requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.7
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
JPype1==0.6.3
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.2
openpyxl==2.5.3
pandas==0.22.0
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
Werkzeug==0.14.1
